==> research/02_data_validation.ipynb <==
import os

%pwd

os.chdir("d://Repositories/ML-geo/")

%pwd

from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class DataValidationConfig:
    root_dir: Path
    STATUS_FILE: str
    STATUS_FILE_DTYPE: str
    unzip_data_dir: Path 
    all_schema: dict

from XGboost_for_slopes.constants import *
from XGboost_for_slopes.utils.common import read_yaml, create_directories

class ConfigurationManager:
    def __init__(self,
                 config_filepath = CONFIG_FILE_PATH,
                 params_filepath = PARAMS_FILE_PATH,
                 schema_filepath = SCHEMA_FILE_PATH):
        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)
        create_directories([self.config.artifacts_root])


    def get_data_validation_config(self) -> DataValidationConfig:
        config = self.config.data_validation
        schema = self.schema.COLUMNS

        create_directories([config.root_dir])

        data_validation_config = DataValidationConfig(
            root_dir=config.root_dir,
            STATUS_FILE= config.STATUS_FILE,
            STATUS_FILE_DTYPE = config.STATUS_FILE_DTYPE,
            unzip_data_dir= config.unzip_data_dir,
            all_schema = schema,
            )
        return data_validation_config

import os 
from XGboost_for_slopes import logger
import pandas as pd

class DataValidation():
    def __init__(self, config: DataValidationConfig):
        self.config = config

    def validate_all_columns(self) -> bool:
        try:
            validation_status = None

            data = pd.read_csv(self.config.unzip_data_dir)
            all_cols = list(data.columns)
            all_schema = self.config.all_schema.keys()

            if all_cols != list(all_schema):
                validation_status = False
                with open(self.config.STATUS_FILE, "w") as f:
                    f.write(f"Validation status: {validation_status}")
                    logger.info(f"Validation status of columns names: {validation_status}")

            else:
                validation_status =  True
                with open(self.config.STATUS_FILE, "w") as f:
                    f.write(f"Validation status: {validation_status}")
                    logger.info(f"Validation status of columns names: {validation_status}")


        except Exception as e:
            raise e
        

    def validate_all_columns_dtypes(self) -> bool:
        try:
            validation_status = None

            data = pd.read_csv(self.config.unzip_data_dir)
            all_cols = list(data.dtypes)
            all_schema = self.config.all_schema.values()


            if all_cols != list(all_schema):
                validation_status = False
                with open(self.config.STATUS_FILE_DTYPE, "w") as f:
                    f.write(f"Validation status: {validation_status}")
                    logger.info(f"Validation status of columns dtypes: {validation_status}")
            else:
                validation_status =  True
                with open(self.config.STATUS_FILE_DTYPE, "w") as f:
                    f.write(f"Validation status: {validation_status}")
                    logger.info(f"Validation status of columns dtypes: {validation_status}")
        except Exception as e:
            raise e


try:
    config = ConfigurationManager()
    data_validation_config = config.get_data_validation_config()
    data_validation = DataValidation(config = data_validation_config)
    data_validation.validate_all_columns()
    data_validation.validate_all_columns_dtypes()

except Exception as e:
    raise e
