==> research/03_data_transformation.ipynb <==
import os

%pwd

os.chdir("d:/Repositories/ML-geo/")

from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class DataTransformationConfig:
    root_dir: Path
    data_path: Path
    transformation_path: Path

from XGboost_for_slopes.constants import *
from XGboost_for_slopes.utils.common import read_yaml, create_directories

class ConfigurationManager:
    def __init__(self,
                 config_filepath = CONFIG_FILE_PATH,
                 params_filepath = PARAMS_FILE_PATH,
                 schema_filepath = SCHEMA_FILE_PATH):
        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)
        create_directories([self.config.artifacts_root])


    def get_data_transformation_config(self) -> DataTransformationConfig:
        config = self.config.data_transformation

        create_directories([config.root_dir])

        data_transformation_config = DataTransformationConfig(
            root_dir=config.root_dir,
            data_path=config.data_path,
            transformation_path= config.transformation_path
            )
        return data_transformation_config

import os 
from XGboost_for_slopes import logger
import pandas as pd
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler,PolynomialFeatures
from sklearn.impute import SimpleImputer
from sklearn.model_selection import GridSearchCV, train_test_split
import dill
from src.XGboost_for_slopes.utils.common import save_object

class DataTransformation():
    def __init__(self, config: DataTransformationConfig):
        self.config = config

    def get_data_transformer_object(self):

        """
        This function is responsible for data transformation
        """
        try:

            num_Pipeline = Pipeline(
                steps=[
                ("imputer", SimpleImputer(strategy="median") ),
                ("scaler", StandardScaler()),
                ("polynomial-scaler", PolynomialFeatures(degree=2))
                ])
            logger.info("Data Transformation file created")

            return num_Pipeline
        except Exception as e:
            raise e

    def get_train_test_data(self):
        data = pd.read_csv(self.config.data_path)
        data = data[data["FSs"] < 3] # Engineers recommendation
        train, test = train_test_split(data,test_size=0.3)
        train.to_csv(os.path.join(self.config.root_dir, "train.csv"), index= False)
        test.to_csv(os.path.join(self.config.root_dir, "test.csv"), index= False)

        logger.info("Splitted data into training and test sets")
        logger.info(train.shape)
        logger.info(test.shape)
        
    def initiate_data_transformation(self):
        try:

            logger.info("Obtaining preprocessing object")

            preprocessing_obj = self.get_data_transformer_object()
            save_object(self.config.transformation_path, preprocessing_obj)
            
        except:
            pass

def save_object(file_path, obj):
    try:
        dir_path= os.path.dirname(file_path)
        os.makedirs(dir_path, exist_ok=True)

        with open(file_path, "wb") as file_obj:
            dill.dump(obj, file_obj)

    except Exception as e:
        raise e
    


try:
    config = ConfigurationManager()
    data_transformation_config = config.get_data_transformation_config()
    data_transformation = DataTransformation(config = data_transformation_config)
    data_transformation.get_train_test_data()
    data_transformation.initiate_data_transformation()

except Exception as e:
    raise e
